Added application configuration as Module.value. This object holds common configuration easily accessible throughout the application. Services can refer to this object in order to retrieve default server's address and port for backend communication.
Added a dumb HTTP interceptor to get acquainted with the whole
concept.

TestInterceptor catches 401 replies from backend, logs them to
the console and broadcast "event:TestInterceptro-ERROR"
which gets caught by AddressbookCtrl and logged again to the console.

This can be used to force client side application, to redirect
to authentication page upon each restricted request, in case
there is no user logged in yet.

Errors other than 401 are bubbled up the chain unfiltered.

Multiple HTTP interceptors can be installed to deal with particular
error codes, or to preprocess requests/responses.

Indeed, preprocessing error responses can be useful to normalize
error information before they get passed to the rest of the application.

This perhaps might be the subject of the next development iteration.